Add experimental children support; fix cancelable events
There exists now a Children field in a Node that can have nested
content. The X, Y, Width, and Height of a parent are passed to children
during layout now and are used to assign a child's position and size as
the outer width/height. Note that the rendering of said children is not
containerized in any way but exist in the global screen space. In the
future we will probably want to have a form of widget
containerization...

Additionally, event generation now uses pointers initially so we can
properly use the EventCancelable interface.

Event propagation is now checked from the deepest child first then
iterating back to the parent nodes from there. This allows a child
widget to cancel an event if it doesn't want the parent to activate (say
for a checkbox within a larger button or similar).