Add Disabled node/widget functionality
As per most things in rebui, the actual effects of this state is largely
widget-defined. However, the Basic widget, which rebui widgets embed,
use this during Hit, so any events will not be sent to a disabled widget.
Potentially we should add a default style for the disabled state. Perhaps
style should have enums(or just strings) for state, such as active,
disabled, hover, default, etc., and the given element receives a state
color assignment to use during rendering...