Replace layout positioning string parsing logic with tokenizer
Instead of doing string checks, regex, etc., we're using a tokenizer and
working from the output of that. This will allow width/height based %
calculations and otherwise.

At the moment we're using a fork of a tokenizer package with my own
changes to allow proper negative number parsing. To be replaced with
proper upstream version once the PR is accepted.